[level 2] Title: k진수에서 소수 개수 구하기, Time: 0.09 ms, Memory: 76.4 MB -BaekjoonHub

+++ 프로그래머스/lv2/92335. k진수에서 소수 개수 구하기/k진수에서 소수 개수 구하기.java
@@ -11,7 +11,7 @@ class Solution {
             long decimalNum = Long.parseLong(kNum);
             
             isPrime = true;
-            for (int i = 2; i <= (int)Math.sqrt(decimalNum); i++){
+            for (int i = 2; i <= Math.sqrt(decimalNum); i++){
                 if (decimalNum % i == 0){
                     isPrime = false;
                 }
